`int32*` to `long*`

+++ Codec/decoder.c
@@ -24305,7 +24305,7 @@ void TransformInverseFrameThreadedToYUV(DECODER *decoder, int frame_index, int n
 										uint8_t *output, int pitch, FRAME_INFO *info,
 										int chroma_offset, int precision)
 {
-	int32_t lPreviousCount,i;
+	long lPreviousCount,i;
 
 	// There are half as many input rows as output rows
 	int transform_height = (((info->height+7)/8)*8) / 2;
@@ -24341,7 +24341,7 @@ void TransformInverseFrameThreadedToRow16u(DECODER *decoder, int frame_index, in
 										   PIXEL16U *output, int pitch, FRAME_INFO *info,
 										   int chroma_offset, int precision)
 {
-	int32_t lPreviousCount,i;
+	long lPreviousCount,i;
 
 	// There are half as many input rows as output rows
 	int transform_height = (((info->height+7)/8)*8) / 2;
